ml-service/resume-nlp/requirements.txt
fastapi
uvicorn
rapidfuzz
python-multipart
pdfminer.six
python-docx
pypdf
